refactor($login): Alteração no login
Foi alterado o local do endpoint e teste
