+++ functions/src/presentation/index.ts
@@ -1,11 +1,7 @@
 import * as functions from "firebase-functions";
 import cors from "cors";
 import * as admin from "firebase-admin";
-import express, { Request, Response } from "express";
-import { UserDataBase } from "../data/userDataBase";
-import { BcryptImplamantation } from "../services/bcryptCryptography";
-import { JWTCryptography } from "../services/JWTCryptography";
-import { LoginUC, LoginUCInput } from "../business/useCases/auth/login";
+import express from "express";
 import { createUserEndpoint } from "./endpoints/createUser";
 import { sendCpfEndpoint } from "./endpoints/sendCpf";
 import { sendFullNameEndpoint } from "./endpoints/sendFullName";
@@ -13,35 +9,14 @@ import { sendAdressEndpoint } from "./endpoints/sendAdress";
 import { sendBirthdayEndpoint } from "./endpoints/sendBirthday";
 import { sendPhoneNumberEndpoint } from "./endpoints/sendPhoneNumber";
 import { sendAmountRequestedEndpoint } from "./endpoints/sendAmountRequested";
+import { loginEndpoint } from "./endpoints/login";
 
 admin.initializeApp();
 const app = express();
 app.use(cors({ origin: true }));
 
 app.post("/createUser", createUserEndpoint);
-
-app.post("/login", async (req: Request, res: Response) => {
-  try {
-    const useCase = new LoginUC(
-      new UserDataBase(),
-      new BcryptImplamantation(),
-      new JWTCryptography()
-    );
-
-    const input: LoginUCInput = {
-      email: req.body.email,
-      password: req.body.password,
-    };
-
-    const result = await useCase.execute(input);
-    res.status(200).send(result);
-  } catch (err) {
-    res.status(400).send({
-      message: err.message,
-    });
-  }
-});
-
+app.post("/login", loginEndpoint);
 app.post("/sendCpfUser", sendCpfEndpoint);
 app.post("/sendFullName", sendFullNameEndpoint);
 app.post("/sendFhoneNumber", sendPhoneNumberEndpoint);
